creating before each method for tests moving test classes to a new package
