+++ src/test/java/test/tests/CourseTest.java
@@ -1,18 +1,26 @@
-package test;
+package test.tests;
 
+import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
+
+import test.Util;
+import test.pageobjects.CoursesPage;
+import test.tests.base.FunctionalTest;
 
 public class CourseTest extends FunctionalTest {
 
 	Util util = new Util();
-
+	private CoursesPage coursesPage;
+	
+	@BeforeMethod
+	protected void navigateToCourses() {
+		//driver.manage().window().maximize();
+		driver.get("http://localhost:3000/course");
+		coursesPage = new CoursesPage(driver);
+	}
+	
 	@Test
 	public void newCourse() {
-
-		driver.manage().window().maximize();
-		driver.get("http://localhost:3000/course");
-
-		CoursesPage coursesPage = new CoursesPage(driver);
 		coursesPage.clickOnAdd();
 		coursesPage.clickOnDeveloperCourseName();
 		coursesPage.enterDeveloperCourseName(util.randomString(10));
@@ -25,11 +33,6 @@ public class CourseTest extends FunctionalTest {
 
 	@Test
 	public void editCourse() {
-
-		driver.manage().window().maximize();
-		driver.get("http://localhost:3000/course");
-
-		CoursesPage coursesPage = new CoursesPage(driver);
 		coursesPage.clickOnAnyData();
 		coursesPage.clickOnDeveloperCourseName();
 		coursesPage.enterDeveloperCourseName(util.randomString(7));
@@ -42,15 +45,9 @@ public class CourseTest extends FunctionalTest {
 
 	@Test
 	public void deleteCourse() {
-
-		driver.manage().window().maximize();
-		driver.get("http://localhost:3000/course");
-
-		CoursesPage coursesPage = new CoursesPage(driver);
 		coursesPage.clickOnAnyData();
 		coursesPage.clickOnDeveloperCourseName();
 		coursesPage.clickOnDelete();
-
 	}
 
 }
